Rename KHR_quantized_geometry to KHR_mesh_quantization
