+++ loaders/src/glTF/2.0/Extensions/KHR_mesh_quantization.ts
